move progress_bar to submodule
