+++ metaseq/logging/progress_bar_dir/progress_bar.py
@@ -20,7 +20,7 @@ from typing import Optional
 
 import torch
 
-from .meters import AverageMeter, StopwatchMeter, TimeMeter
+from metaseq.logging.meters import AverageMeter, StopwatchMeter, TimeMeter
 
 logger = logging.getLogger(__name__)
